+++ GameEngine/src/main/java/gameengine/GameEngine.java
@@ -189,7 +189,7 @@ public class GameEngine implements ApplicationListener {
                 if (image.isRepeat()) {
                     spriteBatch.setProjectionMatrix(camera.combined);
                     spriteBatch.begin();
-                    spriteBatch.draw(animator.getSpellAnimation(), p.getX(), p.getY(), 0, animator.getSpellAnimation().getRegionHeight() / 2, animator.getSpellAnimation().getRegionWidth(), animator.getSpellAnimation().getRegionHeight(), 1, 1, e.getAngle());
+                    spriteBatch.draw(animator.getSpellAnimation(), p.getX(), p.getY(), 0, animator.getSpellAnimation().getRegionHeight() / 2, animator.getSpellAnimation().getRegionWidth(), animator.getSpellAnimation().getRegionHeight(), 1, 1, (float) e.getAngle());
                     spriteBatch.end();
                 }
             }
